+++ tests/test_cli/test_bot_send.py
@@ -240,3 +240,71 @@ class TestVideo:
         )
         result = cli_runner.invoke(cli, args)
         assert result.exit_code == 0
+
+
+class TestAudio:
+    @SKIPIF_HAS_NOT_CONNECTED
+    @SKIPIF_BOT_TOKEN_NOT_EXISTS
+    @SKIPIF_BOT_RECEIVER_NOT_EXISTS
+    def test_vanilla(self, cli_runner: CliRunner, cli, receiver_id: str):
+        args = (
+            "bot",
+            "send",
+            "-r",
+            receiver_id,
+            "audio",
+            "tests/resources/placeholder.wav",
+        )
+        result = cli_runner.invoke(cli, args)
+        assert result.exit_code == 0
+
+    @SKIPIF_HAS_NOT_CONNECTED
+    @SKIPIF_BOT_TOKEN_NOT_EXISTS
+    @SKIPIF_BOT_RECEIVER_NOT_EXISTS
+    def test_performer(self, cli_runner: CliRunner, cli, receiver_id: str):
+        args = (
+            "bot",
+            "send",
+            "-r",
+            receiver_id,
+            "audio",
+            "tests/resources/placeholder.wav",
+            "--performer",
+            "Eray Erdin",
+        )
+        result = cli_runner.invoke(cli, args)
+        assert result.exit_code == 0
+    
+    @SKIPIF_HAS_NOT_CONNECTED
+    @SKIPIF_BOT_TOKEN_NOT_EXISTS
+    @SKIPIF_BOT_RECEIVER_NOT_EXISTS
+    def test_title(self, cli_runner: CliRunner, cli, receiver_id: str):
+        args = (
+            "bot",
+            "send",
+            "-r",
+            receiver_id,
+            "audio",
+            "tests/resources/placeholder.wav",
+            "--title",
+            "White Noise"
+        )
+        result = cli_runner.invoke(cli, args)
+        assert result.exit_code == 0
+
+    @SKIPIF_HAS_NOT_CONNECTED
+    @SKIPIF_BOT_TOKEN_NOT_EXISTS
+    @SKIPIF_BOT_RECEIVER_NOT_EXISTS
+    def test_message(self, cli_runner: CliRunner, cli, receiver_id: str, invoke_message_factory):
+        args = (
+            "bot",
+            "send",
+            "-r",
+            receiver_id,
+            "audio",
+            "tests/resources/placeholder.wav",
+            "-m",
+            invoke_message_factory(self.__class__, self.test_message)
+        )
+        result = cli_runner.invoke(cli, args)
+        assert result.exit_code == 0