+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.4.0a2] - 2019-09-19
+### Changed
+ - The description of `--horizontal` and `--vertical` options on `bot send video`
+ subcommand has been updated to imply that these values are going to be used for
+ determining the aspect ratio of the video thumbnail.
+ - Default values of `--horizontal` and `--vertical` options on `bot send video`
+ has been set to `1`.
+
 ## [v0.4.0a1] - 2019-09-18
 ### Changed
  - Only migrated all subcommands to their packages and modules.